gh-108216: Add pycore_dict_struct.h internal header file
Add a new internal pycore_dict_struct.h header file. It can be used
by debuggers and profilers to inspect a Python dictionary without
having to call Python functions. It should be usable in C++.

pycore_dict.h includes pycore_dict_struct.h (no API change).
